TypeToClimb/backend/requirements.txt
flask==3.1.0
flask-socketio==5.5.1
flask-cors==5.0.1
eventlet==0.40.0
psycopg[binary]==3.3.3
python-dotenv==1.1.0
